pyproject.toml
[tool.poetry]
name = "book-ai-evaluation"
version = "0.1.0"
description = ""
authors = ["Abdur-Rahim-sheikh <abi.rahim.sheikh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^1.1.2"
fonttools = "^4.55.3"
matplotlib = "^3.10.0"
pandas = "^2.2.3"
opencv-python = "^4.10.0.84"
ollama = "^0.4.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
